requirements.txt
python-telegram-bot[job-queue]>=21.0
google-api-python-client
google-auth
python-dotenv
fuzzywuzzy
python-Levenshtein
openai
